// src/cms/payload-types.ts
export interface Media {
  id: number;
  /**
   * Deskripsi singkat isi gambar untuk pembaca layar dan Google.
   */
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}